django-app/apps/diets/diets_api/serializers.py
from rest_framework import serializers
from apps.diets.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ('id', 'name', 'brand')
