django-app/requirements.txt
asgiref==3.5.2
Django==4.1.4
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
python-dotenv==0.20.0
django-crispy-forms==1.14.0
django-bootstrap-v5
django-widget-tweaks
django-filter==23.2
djangorestframework==3.14.0
black==22.12.0